bfloat16 makes offload work

+++ pose_diffusion/models/pointnet2_utils.py
@@ -195,7 +195,8 @@ class PointNetSetAbstraction(nn.Module):
         new_points = new_points.permute(0, 3, 2, 1) # [B, C+D, nsample,npoint]
         for i, conv in enumerate(self.mlp_convs):
             bn = self.mlp_bns[i]
-            new_points = new_points.float()
+            # set new_points to be float16
+            new_points = new_points.bfloat16()
             new_points =  F.relu(bn(conv(new_points)))
 
         new_points = torch.max(new_points, 2)[0]
@@ -312,7 +313,7 @@ class PointNetFeaturePropagation(nn.Module):
         new_points = new_points.permute(0, 2, 1)
         for i, conv in enumerate(self.mlp_convs):
             bn = self.mlp_bns[i]
-            new_points = new_points.float()
+            new_points = new_points.bfloat16()
             new_points = F.relu(bn(conv(new_points)))
         return new_points
